rust/host: prove the RLS fence, not this crate, refuses a stale-era write
Fixes a real, latent type bug the new test caught immediately: era/
ordinal columns are Postgres `int` (int4) in Ruby's actual DDL
(era_store.rb's `ordinal int NOT NULL`, provisioning.rb's `era int NOT
NULL`) -- LineageConfig::era was i64 everywhere, which tokio_postgres
rejects outright ("cannot convert between the Rust type i64 and the
Postgres type int4") the moment it hits a real Ruby-provisioned table.
The existing dispatch.rs test fixture had quietly worked around this by
declaring its OWN mock schema as bigint instead of int, which is why it
never caught the mismatch -- fixed to match Ruby's real column type
exactly instead of a convenient stand-in for it. LineageConfig::era,
era_exists, head_snapshot_table, and HECKS_ERA parsing are now i32
throughout.

New test (journal.rs): a Ruby fixture script
(tests/fixtures/mint_stale_era.rb, a close port of
spec/adapters/driven/postgres/lineage_spec.rb's own "fences a
deployment's app role" setup -- proven Ruby code, not reinvented) mints
a REAL era 2 via LineageManager against a scratch database with a
genuinely fenced non-superuser app role. The Rust test then connects AS
that role and proves append_lineage_mutation succeeds under era 2 and
fails under era 1 with Postgres's own "row-level security policy"
message -- the central claim journal.rs's own header makes (staleness
is Postgres's job, not a Rust-side check) demonstrated, not just
argued. All 5 rust/host tests pass against real Postgres.

+++ rust/host/src/main.rs
@@ -67,7 +67,12 @@ async fn main() -> Result<(), Error> {
     // under the wrong domain or era is exactly the failure mode this
     // whole design exists to prevent.
     let domain = std::env::var("HECKS_DOMAIN").map_err(|_| "HECKS_DOMAIN is required")?;
-    let era: i64 = std::env::var("HECKS_ERA")
+    // i32, matching Postgres `int` -- hecks_eras.ordinal and every
+    // journal's own `era` column are `int` (int4) in Ruby's real DDL
+    // (era_store.rb/provisioning.rb), not bigint; tokio_postgres
+    // requires the Rust and Postgres types to match exactly, not just
+    // be numerically compatible.
+    let era: i32 = std::env::var("HECKS_ERA")
         .map_err(|_| "HECKS_ERA is required")?
         .parse()
         .map_err(|e| format!("HECKS_ERA must be an integer era ordinal: {e}"))?;
